feat: type에 따른 스타일 변경
-game or session 따른 스타일 변경을 위한 varaints파일 및 구조 변경
