+++ src/app/(auth)/(with-header)/correctAnswerRate/correctAnswerRate.type.ts
@@ -1,0 +1,4 @@
+export interface CorrectAnswerRateProps {
+  averageScore: number;
+  type: "game" | "session";
+}
