+++ src/app/(auth)/(with-header)/correctAnswerRate/index.ts
@@ -1,0 +1,2 @@
+export * from "./CorrectAnswerRate";
+export * from "./correctAnswerRate.type";
